+++ src/phone/phone_instance_t.cpp
@@ -377,7 +377,7 @@ void phone_instance_t::adjust_rx_level_for_capture_device(float level) const {
     }
 }
 
-unsigned int phone_instance_t::get_tx_level_adjustment_for_capture_device() const {
+float phone_instance_t::get_tx_level_adjustment_for_capture_device() const {
     try {
         return get_level_adjustment_for_capture_device(phone::tx_rx_direction::Transmit);
     } catch (const pj::Error& e) {
@@ -385,7 +385,7 @@ unsigned int phone_instance_t::get_tx_level_adjustment_for_capture_device() cons
     }
 }
 
-unsigned int phone_instance_t::get_rx_level_adjustment_for_capture_device() const {
+float phone_instance_t::get_rx_level_adjustment_for_capture_device() const {
     try {
         return get_level_adjustment_for_capture_device(phone::tx_rx_direction::Receive);
     } catch (const pj::Error& e) {
@@ -393,7 +393,7 @@ unsigned int phone_instance_t::get_rx_level_adjustment_for_capture_device() cons
     }
 }
 
-unsigned int phone_instance_t::get_level_adjustment_for_capture_device(phone::tx_rx_direction direction) const {
+float phone_instance_t::get_level_adjustment_for_capture_device(phone::tx_rx_direction direction) const {
     auto portInfo = m_ep->audDevManager().getCaptureDevMedia().getPortInfo();
     switch (direction) {
         case phone::tx_rx_direction::Transmit:
